requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bioc==2.0.post5
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
cssselect==1.1.0
Cython==0.29.28
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
docopt==0.6.2
eutils==0.6.0
filelock==3.7.1
gensim==4.2.0
habanero==1.2.2
html5lib==1.1
huggingface-hub==0.12.0
humanfriendly==10.0
idna==3.3
iniconfig==1.1.1
intervaltree==3.1.0
jaro-winkler==2.0.0
joblib==1.2.0
jsonlines==3.1.0
lxml==4.9.0
metapub==0.5.5
nltk==3.8.1
numpy==1.22.4
orcid==1.0.3
packaging==21.3
pandas==1.4.2
patsy==0.5.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyMuPDF==1.19.6
pyparsing==3.0.9
pyreadline3==3.4.1
pytest==7.1.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
PyYAML==6.0
regex==2022.10.31
requests==2.27.1
scikit-learn==1.2.1
scipy==1.8.1
semanticscholar==0.4.0
simplejson==3.19.1
six==1.16.0
smart-open==6.0.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.2
statsmodels==0.13.5
tabulate==0.8.9
tenacity==8.1.0
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tox==3.25.0
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
